+++ source/extensions/isaacsim.simulation_app/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.17.0] - 2026-03-23
+### Added
+- Emit telemetry event for app startup duration via isaacsim.core.telemetry
+### Changed
+- Updated return types
+
 ## [2.16.1] - 2026-03-07
 ### Changed
 - Automatically close the application during interpreter shutdown if close() was not called